+++ CHANGELOG.rst
@@ -3,6 +3,14 @@ Version explained:
  - minor : increase on risk of code breakage during a major release
  - bug   : increase on bug or incremental changes
 
+Version 4.2.22
+ * Fix: route reload for offline neighbors #1126
+   patch: Malcolm Dodds
+ * Fix: make sure we compare next-hop self and next-hop IP correctly (#1153)
+   reported: gitneep
+ * Compatibility: remove "not-a-fragment" "!is-fragment" should be used instead
+ * Upgrade six to the latest version
+
 Version 4.2.21
  * Fix: regressing on announcing routes from the API #1108